dynamic_pricing/requirements.txt
# Dynamic Pricing with PPO - Requirements
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
requests>=2.25.0
flask>=2.0.0
flask-cors>=3.0.10
gymnasium>=0.28.0
stable-baselines3>=2.0.0
tensorboard>=2.10.0
joblib>=1.1.0
tqdm>=4.64.0
plotly>=5.0.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0 